+++ app/features/movie/apis/index.ts
@@ -1,8 +1,8 @@
-export * from "./getMovieDetail";
-export * from "./getMovieTrailer";
-export * from "./getMoviesByGenre";
-export * from "./getMoviesByNowPlaying";
-export * from "./getMoviesByTopRated";
-export * from "./getMoviesByUpcoming";
-export * from "./saveMovie";
-export * from "./toggleLikeMovie";
+export * from "./get-movie-detail-info";
+export * from "./get-movie-trailer";
+export * from "./get-movies-by-genre";
+export * from "./get-movies-by-now-playing";
+export * from "./get-movies-by-top-rated";
+export * from "./get-movies-by-upcoming";
+export * from "./save-movie-info";
+export * from "./toggle-like-movie";
