+++ app/features/review/apis/index.ts
@@ -1,9 +1,9 @@
 export * from "./deleteReview";
-export * from "./toggleLikeReview";
-export * from "./postNewReview";
-export * from "./updateReview";
-export * from "./getMyReviewsByMovie";
-export * from "./getReviewsByMovie";
-export * from "./getRecentReviews";
-export * from "./getReviewsByMovieLimit";
-export * from "./getMyReviewsToMyPage";
+export * from "./toggle-like-review";
+export * from "./post-new-review";
+export * from "./update-review";
+export * from "./get-my-review-by-movie";
+export * from "./get-all-reviews-by-movie";
+export * from "./get-recent-reviews-to-main";
+export * from "./get-reviews-by-movie-limit";
+export * from "./get-my-reviews-to-mypage";
